Update hook constructors and fix issues with import order.
