+++ virtool_workflow/__init__.py
@@ -1,18 +1,9 @@
 """An SDK for defining Virtool workflows."""
-from virtool_workflow.fixtures import FixtureScope, fixture
-from virtool_workflow.execution import hooks
-from virtool_workflow.execution.hooks.hooks import hook, Hook
-from virtool_workflow.execution.workflow_execution import WorkflowExecution, State, WorkflowError
-from virtool_workflow.workflow import Workflow
 from virtool_workflow.decorator_api import step, cleanup, startup
+from virtool_workflow.fixtures import FixtureScope, fixture
+from virtool_workflow.workflow import Workflow
 
 __all__ = [
-    "hooks",
-    "hook",
-    "Hook",
-    "WorkflowExecution",
-    "WorkflowError",
-    "State",
     "FixtureScope",
     "fixture",
     "Workflow",
