+++ virtool_workflow/execution/hooks/hooks.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 class Hook:
 
-    def __init__(self, hook_name, parameters, return_type):
+    def __init__(self, hook_name):
         """
         A set of functions to be called as a group upon a particular event.
 
@@ -18,15 +18,8 @@ class Hook:
         are validated to match the types provided.
 
         :param hook_name: The name of this hook.
-        :param parameters: A list of types for the parameters a callback function
-            should accept. These will be used to validate function signatures before
-            adding them to the set.
-
-        :param return_type: The expected return type for callback functions.
         """
         self.name = hook_name
-        self._params = parameters
-        self._return = return_type
 
         self.callbacks = []
